[Refactor] refactor symm_mempory code
