+++ magi_compiler/passes/fsdp_overlap/symm_ag_rewrite.py
@@ -48,7 +48,7 @@ def _input_is_arena_shard_from_local(src: fx.Node) -> bool:
 
 def rewrite_weight_ag_to_copy_engine(graph: fx.GraphModule) -> int:
     """Retarget marked weight gathers to ``magi::symm_all_gather``. Returns count rewritten."""
-    from magi_compiler.runtime.symm_all_gather import SYMM_ALL_GATHER, SYMM_ALL_GATHER_COALESCED
+    from magi_compiler.symm_mem.all_gather import SYMM_ALL_GATHER, SYMM_ALL_GATHER_COALESCED
 
     rewritten = 0
     skipped = 0
